move init logic to dedicated file
